sd: Add preview of diffuser steps (#105)
* sd.cpp: pass output_path always as std::string&

To simplify.

* sd.cpp: add optional addition to filename in save_image()

For example, image.png.preview.png.
This is needed for preview suffix, to append same extension as in base filename.

* sd.cpp: add option to decode diffuser steps

* sd.cpp: add option to preview diffuser output fast

* sd.cpp: add option for scaled previews

* Description in ReadMe

* sd.cpp: better preview variant

* sd.cpp: more preview simplification

* Manual expansion of channels cycle

Since number of channels is low, it's probably acceptable and makes smaller binary than -O3.

* sd.cpp: add preview timings

* sd.cpp: show time as integer

If operation takes longer than 1000 seconds, the time is displayed as "1.21752e+06ms" (instead of 1217520ms), this is probably less convenient.